Refactor CLI and agent core into package
