+++ agent/__init__.py
@@ -1,0 +1,4 @@
+"""Core assembly of the Universal AI agent."""
+from .core import handle_query, agent_executor, ResearchResponse, parser
+
+__all__ = ["handle_query", "agent_executor", "ResearchResponse", "parser"]
